refactor(python)!: drop bloqade.gemini.device re-exports from bloqade.lanes
bloqade.lanes re-exported GeminiLogicalSimulator / GeminiLogicalSimulatorTask
/ DetectorResult / Result from bloqade.gemini.device, making the low-level
lanes package depend up on the high-level gemini package while gemini depends
down on lanes everywhere. That cycle forced a fragile TYPE_CHECKING +
__getattr__ lazy-export shim (it had even produced a captured ImportError in
demo/simulator_device_demo.ipynb). Remove the re-exports so the up-dependency
and that part of the shim are gone; the canonical import is bloqade.gemini.

Migrated the few callers (a regression test + two demos + a notebook) to
`from bloqade.gemini import GeminiLogicalSimulator`. The lanes-internal
exports (Metrics, noise-model helpers, steane defaults) stay lazily exported
because deeper lanes submodules still transitively import gemini.

BREAKING CHANGE: `from bloqade.lanes import GeminiLogicalSimulator` (and
GeminiLogicalSimulatorTask / DetectorResult / Result) no longer works; import
them from `bloqade.gemini` instead.
